Wrap with a TestToolsGesture detector on dev only
